docker-service/python_reddit/python_requirements.txt
praw
pandas 
numpy
confluent-kafka
cassandra-driver